Add initial cf-manifest generator.
This doesn't actually generate manifests -- I want to discuss what to
generate with @LinuxBozo and the rest of Engineering. In the meantime, this
generates language-specific Procfiles and a TODO entry.

I assume we'll want to define full manifests for each language and try to fill
the TODO entry with a user prompt. We'll also likely need to retrieve a
language version from the user.